+++ src/app/HISTORY.md
@@ -15,6 +15,13 @@ App Framework Change History
 * ModelSync.REST's `serialize()` method now receives the `action` which the
   `sync()` method was invoked with. [Ticket #2532625]
 
+### ModelList
+
+* You may now add models to a ModelList at instantiation time by providing an
+  Object, array of Objects, Model instance, array of Model instances, or another
+  ModelList instance in the `items` property of the config object passed to
+  ModelList's constructor. This change also applies to LazyModelList.
+
 
 3.6.0
 -----